requirements.txt
amqp==5.3.1
asgiref==3.8.1
asyncio==3.4.3
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
Flask==3.1.0
flask-cors==5.0.1
googleapis-common-protos==1.65.0
grpclib==0.4.7
h2==4.1.0
haversine==2.9.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
kombu==5.5.0
libmagic==1.0
MarkupSafe==3.0.2
memory_profiler==0.60.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
protobuf==5.28.2
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pyotp==2.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-magic==0.4.27
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
viam-sdk==0.31.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
